### 2022-08-04 #4648 @vlad0x00
Hi @arfon! It looks like there was a typo in the authors section, I pushed a fix to the `paper` branch.
