+++ include/btllib/mi_bloom_filter.hpp
@@ -24,6 +24,22 @@ public:
 
   static const unsigned BLOCKSIZE = 512;
 
+  /// @cond HIDDEN_SYMBOLS
+#pragma pack(1) // to maintain consistent values across platforms
+  struct FileHeader
+  {
+    char magic[8]; // NOLINT
+    uint32_t hlen; // header length (including spaced seeds)
+    uint64_t size;
+    uint32_t nhash;
+    uint32_t kmer;
+    uint32_t version;
+    uint32_t bucket_size;
+    uint32_t ID_size;
+    uint8_t allowed_miss;
+  };
+  /// @endcond
+
   /** Construct a dummy multi-indexed Bloom filter (e.g. as a default argument).
    */
   MIBloomFilter() {}
@@ -31,10 +47,17 @@ public:
   /**
    * Construct an empty Bloom filter of given size.
    *
-   * @param bytes Filter size in bytes.
+   * @param bv_size Filter size in bv_size.
    * @param hash_fn Name of the hash function used. Used for metadata. Optional.
    */
-  MIBloomFilter(size_t bytes, unsigned hash_num, std::string hash_fn = "");
+  MIBloomFilter(size_t bv_size, unsigned hash_num, std::string hash_fn = "");
+
+  /**
+  * Load a Bloom filter from a file.
+  *
+  * @param path Filepath to load from.
+  */
+  explicit MIBloomFilter(const std::string& path);
 
   /**
    * Transform bit vector to interleaved bit vector and create ID array of size
@@ -111,7 +134,14 @@ public:
   * @param hashes Integer vector of hash values.
   */
   std::vector<T> get_id(const std::vector<uint64_t>& hashes) { return get_id(hashes.data());}
-  
+ 
+  /**
+  * Save the Bloom filter to a file that can be loaded in the future.
+  *
+  * @param path Filepath to store filter at.
+  */
+  void save(const std::string& path);
+ 
   /** Get population count, i.e. the number of 1 bits in the filter. */
   uint64_t get_pop_cnt();
 
@@ -122,8 +152,8 @@ private:
   T get_data(uint64_t rank) const { return id_array[rank]; }
   void set_data(uint64_t pos, T id);
 
-  size_t bytes = 0;
-  unsigned k = 0;
+  size_t bv_size = 0;
+  unsigned kmer_size = 0;
   unsigned hash_num;
   std::string hash_fn;
 
@@ -134,17 +164,111 @@ private:
   T* id_array;
 
   bool bv_insertion_completed = false, id_insertion_completed = false;
+  static const uint32_t MI_BLOOM_FILTER_VERSION = 1;
 };
 
 template<typename T>
-inline MIBloomFilter<T>::MIBloomFilter(size_t bytes,
+inline MIBloomFilter<T>::MIBloomFilter(size_t bv_size,
                                        unsigned hash_num,
                                        std::string hash_fn)
-  : bytes(bytes)
+  : bv_size(bv_size)
   , hash_num(hash_num)
   , hash_fn(hash_fn)
 {
-  bit_vector = sdsl::bit_vector(bytes);
+  bit_vector = sdsl::bit_vector(bv_size);
+}
+
+template<typename T>
+MIBloomFilter<T>::MIBloomFilter(const std::string& filter_file_path)
+     // TODO: make more streamlined
+  {
+    std::cout << "-----xololo1" << std::flush;
+#pragma omp parallel for default(none) shared(filter_file_path)
+    for (unsigned i = 0; i < 2; ++i) {
+      if (i == 0) {
+        FILE* file = fopen(filter_file_path.c_str(), "rbe");
+        check_error(file == nullptr,
+                    "MIBloomFilter: File " + filter_file_path +
+                      " could not be read.");
+
+        FileHeader header;
[…]
