### 2022-08-24 bfdb0c0
indexlr more fixes
