### 2022-08-26 042cf9b
get saturated pop cnt added
