+++ tests/mi_bloom_filter.cpp
@@ -84,6 +84,7 @@ main()
 	TEST_ASSERT(count < expected_id_count + (expected_id_count * tolerance));
 	TEST_ASSERT(count > expected_id_count - (expected_id_count * tolerance));
   }
+  TEST_ASSERT(mi_bf_2.get_pop_saturated_cnt() == 0); // testing no saturation
   std::cerr << "Testing multi-indexed BloomFilter random sampling successful" << std::endl;
   
  
@@ -174,7 +175,7 @@ main()
 		|| (!ID_3_found && (ID_4_found && ID_1_found && ID_2_found))
 		|| (!ID_4_found && (ID_1_found && ID_2_found && ID_3_found))
   );
-  TEST_ASSERT(mi_bf_4.get_pop_saturated_cnt());
+  TEST_ASSERT(mi_bf_4.get_pop_saturated_cnt() > 0);
   std::cerr << "multi-indexed BloomFilter saturation test success." << std::endl;
 
   // TODO: Test MIBloomFilter(sdsl::bit_vector& bit_vector, unsigned hash_num, std::string hash_fn = "");  
