### 2022-08-28 #4648 @arfon
:wave: @fboehm – would you be willing to edit this submission for JOSS? I realize it's a little off-topic for you but I'm hoping you might be able to take it :-)
