### 2022-08-30 #4648 @fboehm
@gtonkinhill, @cmorris2945, @pjb7687, @MaximLippeveld, @VeroniqueLegrand - would you be willing to review this submission for JOSS? The reviews are driven by checklists. Please let me know if you have any questions. Thanks!
