### 2022-08-31 #4720 @editorialbot
[REVIEW]: btllib: A C++ library with Python interface for efficient genomic sequence processing
**Submitting author:** @vlad0x00 (<a href="http://orcid.org/0000-0002-2992-9935">Vladimir Nikolic</a>)
**Repository:** https://github.com/bcgsc/btllib
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.4.9
