### 2022-08-31 5ff5a97
ID count feature added, corresponding test added
