### 2022-09-05 #4720 @MaximLippeveld
I have finished my initial review. I like the overall quality of the software and the paper :+1: I did not check the 'reproducibility' and 'performance' boxes yet.

Concerning the reproducibility, I was not able to find how to reproduce figure 1A and 1B from the paper. Are there any instructions I missed?

Concerning the performance claims, it is stated that ntHash implementation is "A very efficient DNA/RNA rolling hash function, an order of magnitude faster than the best performing alternatives in typical use cases". No evidence seems to be given to support this claim.
