### 2022-09-05 #4720 @pjb7687
I have also finished my initial review. For me, the library seems to be a dedicated library for internal use rather than a general bioinformatics library. But I think the code quality is overall very good and the source code is well maintained on Github. The package will be useful for some use cases, e.g. sequencing data I/O or some algorithms based on bloom filters. There could be some other groups interested in using the package, at least thanks to its fast I/O speed.

However, there are several checkboxes I didn't tick. Below I explain the concerns.

1. One of the main improvements that the authors claimed is the Python wrapper, but there seems to be no Python API documentation. (Functionality documentation)
2. Not all functions seem to be exported to Python. For example, I can't find "CountBloomFilter" in Python library. (Functionality)
3. The authors claim that `btllib` is faster than Heng Li's `kseq` (Figure panel B). That's nice, would be nice if the authors provide the code used to test it. (Reproducibility and Performance)

I also have a minor concern:

1. There are only 3 examples available. It would be really nice if the authors add more.
