### 2022-09-06 ab6e269
comments and fixes for PR
