+++ include/btllib/mi_bloom_filter.hpp
@@ -34,9 +34,6 @@ public:
     uint32_t nhash;
     uint32_t kmer;
     uint32_t version;
-    uint32_t bucket_size;
-    uint32_t ID_size;
-    uint8_t allowed_miss;
   };
   /// @endcond
 
@@ -45,26 +42,30 @@ public:
   MIBloomFilter() {}
 
   /**
-   * Construct an empty Bloom filter of given size.
+   * Construct an empty multi-indexed Bloom filter of given size.
    *
    * @param bv_size Filter size in bv_size.
+   * @param hash_num Number of hash functions to be used.
    * @param hash_fn Name of the hash function used. Used for metadata. Optional.
    */
   MIBloomFilter(size_t bv_size, unsigned hash_num, std::string hash_fn = "");
 
   /**
-  * Construct a Bloom filter with a prebuilt interleaved bit vector.
-  *
-  * @param bv_size Filter size in bv_size.
-  * @param hash_fn Name of the hash function used. Used for metadata. Optional.
-  */
-  MIBloomFilter(sdsl::bit_vector& bit_vector, unsigned hash_num, std::string hash_fn = "");
+   * Construct a multi-indexed Bloom filter with a prebuilt interleaved bit
+   * vector.
+   *
+   * @param hash_num Number of hash functions to be used.
+   * @param hash_fn Name of the hash function used. Used for metadata. Optional.
+   */
+  MIBloomFilter(sdsl::bit_vector& bit_vector,
+                unsigned hash_num,
+                std::string hash_fn = "");
 
   /**
-  * Load a Bloom filter from a file.
-  *
-  * @param path Filepath to load from.
-  */
+   * Load a multi-indexed Bloom filter from a file.
+   *
+   * @param path Filepath to load from.
+   */
   explicit MIBloomFilter(const std::string& path);
 
   /**
@@ -72,13 +73,21 @@ public:
    * of interleaved bit vector. It can be called only once and bit vector
    * insertions cannot be made after calling this function.
    */
-  bool complete_bv_insertion();
+  void complete_bv_insertion();
+
+  /**
+   * Required to pass to saturation stage.
+   */
+  void complete_id_insertion()
+  {
+    id_insertion_completed = true;
+  }
 
   /**
    * Insert an element's hash values to bit vector.
    *
    * @param hashes Integer array of hash values. Array size should equal the
-   * hash_num argument used when the Bloom filter was constructed.
+   * hash_num argument used when the multi-indexed Bloom filter was constructed.
    */
   void insert_bv(const uint64_t* hashes);
 
@@ -86,7 +95,7 @@ public:
    * Insert an element's hash values to bit vector.
    *
    * @param hashes Integer vector of hash values. Array size should equal the
-   * hash_num argument used when the Bloom filter was constructed.
+   * hash_num argument used when the multi-indexed Bloom filter was constructed.
    */
   void insert_bv(const std::vector<uint64_t>& hashes)
   {
@@ -98,7 +107,7 @@ public:
    * This function can run after bit vector insertion is completed.
    *
    * @param hashes Integer array of hash values. Array size should equal the
-   * hash_num argument used when the Bloom filter was constructed.
+   * hash_num argument used when the multi-indexed Bloom filter was constructed.
    */
   bool bv_contains(const uint64_t* hashes);
 
@@ -107,7 +116,7 @@ public:
    * This function can run after bit vector insertion is completed.
    *
    * @param hashes Integer vector of hash values. Array size should equal the
-   * hash_num argument used when the Bloom filter was constructed.
[…]
