+++ tests/mi_bloom_filter.cpp
@@ -3,6 +3,7 @@
 
 #include "helpers.hpp"
 
+#include <cstdio>
 #include <cmath>
 #include <algorithm>
 #include <cstdio>
@@ -13,54 +14,35 @@ int
 main()
 {
   std::cerr << "Testing multi-indexed BloomFilter" << std::endl;
-  btllib::MIBloomFilter<uint8_t> mi_bf(1024 * 1024, 3, "ntHash");
-  mi_bf.insert_bv({ 1, 10, 100 });
-  mi_bf.insert_bv({ 100, 200, 300 });
-  mi_bf.complete_bv_insertion();
+  btllib::MIBloomFilter<uint8_t> mi_bf_1(1024 * 1024, 3, "ntHash");
+  mi_bf_1.insert_bv({ 1, 10, 100 });
+  mi_bf_1.insert_bv({ 100, 200, 300 });
+  mi_bf_1.complete_bv_insertion();
 
-  TEST_ASSERT(mi_bf.bv_contains({ 1, 10, 100 }));
-  TEST_ASSERT(mi_bf.bv_contains({ 100, 200, 300 }));
-  TEST_ASSERT(!mi_bf.bv_contains({ 1, 20, 100 }));
+  TEST_ASSERT(mi_bf_1.bv_contains({ 1, 10, 100 }));
+  TEST_ASSERT(mi_bf_1.bv_contains({ 100, 200, 300 }));
+  TEST_ASSERT(!mi_bf_1.bv_contains({ 1, 20, 100 }));
 
   uint8_t ID_1 = 12;
-  mi_bf.insert_id({ 1, 10, 100 }, ID_1);
+  mi_bf_1.insert_id({ 1, 10, 100 }, ID_1);
   
-  std::vector<uint8_t> results = mi_bf.get_id({ 1, 10, 100 });
-  for(auto& id : results) {
+  std::vector<uint8_t> results_1 = mi_bf_1.get_id({ 1, 10, 100 });
+  for(auto& id : results_1) {
   	TEST_ASSERT_EQ(id, ID_1);
   }
-  std::cerr << "multi-indexed BloomFilter insertion successful" << std::endl;
-  
+
+  std::cerr << "multi-indexed BloomFilter ID count test" << std::endl;
   std::cerr << "Testing ID counting" << std::endl;
   bool include_saturated = true;
-  TEST_ASSERT(mi_bf.get_ID_occurence_count(include_saturated)[ID_1] == 3)
-  std::cerr << "Testing ID counting successfull" << std::endl;
+
+  TEST_ASSERT(mi_bf_1.get_ID_occurence_count(include_saturated)[ID_1] == 3)
+
 
   std::cerr << "Testing multi-indexed BloomFilter random sampling" << std::endl;  
-  std::string random_dna = "";
   int dna_length = 100000;
+  std::string random_dna = get_random_seq(dna_length);
   int expected_id_count = dna_length / 4;
   double tolerance = 0.1;
-  for(int i = 0; i < dna_length; i++){
-	int rnd = rand() % 4;
-	switch(rnd) {
-  	case 0:
-    		random_dna += "A";
-		break;
-        case 1:
-                random_dna += "T";
-		break;
-        case 2:
-                random_dna += "G";
-		break;
-        case 3:
-                random_dna += "C";
-		break;
-        default:
-                break;
-  	}
-  }
-
 
   int counter = 0;
   btllib::MIBloomFilter<uint8_t> mi_bf_2(256 * 1024 * 1024, 1, "ntHash");
@@ -88,18 +70,13 @@ main()
 	TEST_ASSERT(count > expected_id_count - (expected_id_count * tolerance));
   }
[…]
