### 2022-09-06 545cd21
counts array change to comply with C++ standards
