### 2022-09-09 #4720 @vlad0x00
Thank you @MaximLippeveld  and @pjb7687 for your reviews! Here are my responses to your concerns:

> Concerning the reproducibility, I was not able to find how to reproduce figure 1A and 1B from the paper. Are there any instructions I missed?

I am currently in the process of uploading the necessary code to reproduce the figures, I'll post here once available in the repo.

> Concerning the performance claims, it is stated that ntHash implementation is "A very efficient DNA/RNA rolling hash function, an order of magnitude faster than the best performing alternatives in typical use cases". No evidence seems to be given to support this claim.

The claim is from the ntHash publication (https://academic.oup.com/bioinformatics/article/32/22/3492/2525588), which has been cited in the btllib manuscript. The btllib library currently has the latest implementation of the ntHash function. This hash function is still actively developed and there have been further speed improvements in the latest publication: https://academic.oup.com/bioinformatics/advance-article/doi/10.1093/bioinformatics/btac564/6674501?login=false

> One of the main improvements that the authors claimed is the Python wrapper, but there seems to be no Python API documentation. (Functionality documentation)

In the repo README, we state: "The wrappers correspond one-to-one with C++ code so any functions and classes can be used under the same name. The only exceptions are nested classes which are prefixed with outer class name (e.g. btllib::SeqReader::Flag in C++ versus btllib.SeqReaderFlag in Python), and (Kmer)CountingBloomFilter which provides CountingBloomFilter8, CountingBloomFilter16, CountingBloomFilter32, KmerCountingBloomFilter8, KmerCountingBloomFilter16, CountingBloomFilter32 with counters 8, 16, and 32 bits wide."
So the idea is to have the same functionality under the same names as in C++, and this has guided the naming of classes and functions in the library.

> Not all functions seem to be exported to Python. For example, I can't find "CountingBloomFilter" in Python library. (Functionality)

Thank you for spotting this. I looked into it and noticed counting Bloom filters weren't properly wrapped. This has been fixed in the latest (v1.4.5) version).

> The authors claim that btllib is faster than Heng Li's kseq (Figure panel B). That's nice, would be nice if the authors provide the code used to test it. (Reproducibility and Performance)

I will be uploading code to reproduce the figures in the manuscript. Once there, I'll make another comment.
