### 2022-09-13 #4648 @VeroniqueLegrand
Hello, Sorry for the late delay.
Yes, I would be happy to review this software.

Best regards,
Véronique

--
Véronique Legrand
Software & IT engineer – HPC/HTC
https://research.pasteur.fr/en/member/veronique-legrand/
HPC platform
Institut Pasteur Paris
Tel : 01 44 38 95 03
        07 88 15 34 68

From: Frederick Boehm ***@***.***>
Reply-To: openjournals/joss-reviews ***@***.***>
Date: Tuesday 30 August 2022 at 19:08
Cc: Véronique LEGRAND ***@***.***>, Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [PRE REVIEW]: btllib: A C++ library with Python interface for efficient genomic sequence processing (Issue #4648)

@gtonkinhill<https://urldefense.com/v3/__https:/github.com/gtonkinhill__;!!JFdNOqOXpB6UZW0!pJavYWIH1eV8W7b2E4od-nKDk2RljfYAKEGOB_0x5qYaS6kONFgIUiP26r3K-DkG9DkSAztkyPDkE4sIg3TfvuJ8iQI$>, @cmorris2945<https://urldefense.com/v3/__https:/github.com/cmorris2945__;!!JFdNOqOXpB6UZW0!pJavYWIH1eV8W7b2E4od-nKDk2RljfYAKEGOB_0x5qYaS6kONFgIUiP26r3K-DkG9DkSAztkyPDkE4sIg3TfTG49roU$>, @pjb7687<https://urldefense.com/v3/__https:/github.com/pjb7687__;!!JFdNOqOXpB6UZW0!pJavYWIH1eV8W7b2E4od-nKDk2RljfYAKEGOB_0x5qYaS6kONFgIUiP26r3K-DkG9DkSAztkyPDkE4sIg3TfSENYzWo$>, @MaximLippeveld<https://urldefense.com/v3/__https:/github.com/MaximLippeveld__;!!JFdNOqOXpB6UZW0!pJavYWIH1eV8W7b2E4od-nKDk2RljfYAKEGOB_0x5qYaS6kONFgIUiP26r3K-DkG9DkSAztkyPDkE4sIg3TfR-vIKIw$>, @VeroniqueLegrand<https://urldefense.com/v3/__https:/github.com/VeroniqueLegrand__;!!JFdNOqOXpB6UZW0!pJavYWIH1eV8W7b2E4od-nKDk2RljfYAKEGOB_0x5qYaS6kONFgIUiP26r3K-DkG9DkSAztkyPDkE4sIg3Tfu95f6is$> - would you be willing to review this submission for JOSS? The reviews are driven by checklists. Please let me know if you have any questions. Thanks!
