### 2022-09-15 1fd2af7
Add benchmark scripts
