+++ paper.md
@@ -25,7 +25,7 @@ authors:
     orcid: 0000-0002-6875-4939
   - name: Justin Chu
     affiliation: "1, 2"
-  - name: René Warren
+  - name: René L. Warren
     affiliation: 1
     orcid: 0000-0002-9890-2293
   - name: Inanç Birol
