### 2022-09-16 723dfa3
Update figure
