+++ paper.md
@@ -63,7 +63,7 @@ The library has the implementation of the following algorithms and data structur
 - **Sequence I/O** SeqReader and SeqWriter classes provide efficient and flexible I/O for sequence files. SeqReader is capable of reading sequences in different formats such as FASTA and FASTQ including multiline, and SAM format. The format is automatically detected even without file extension or if the data is piped. SeqReader also supports multiple threads to read in parallel, each thread receiving a copy of the sequence that can be modified as well as ad-hoc compression and decompression of the data in common formats (gzip, bzip2, xz, lrzip, zip, 7zip). \autoref{fig:scalability} B) shows the scalability of using multiple threads to load and process sequences.
 - **Utility functions** Various functions for common tasks such as reverse complementation, string manipulation, and logging.
 
-![**A)** The average counter difference between 10GiB Counting Bloom Filters with single threaded and multithreaded insertion of k-mers of 250,000 long reads. The single threaded version does not suffer from race conditions and thus has the ground truth. The more threads we add the more visible is the effect of race conditions, but thanks to the mitigation mechanism, the differences are small. **B)** Wall-clock time it takes to read in and run a simulated workload of 5ms per read on 250,000 long reads. The blue data points use efficient sequence reading code, kseq [@kseq], surrounded by a naive implementation of a critical section,  while the orange data points use SeqReader. Unlike the critical section implementation which plateaus after a number of threads, SeqReader scales well and keeps benefitting even from higher number of threads.\label{fig:scalability}](scalability.png){ width=100% }
+![**A)** The average counter difference between 3GiB Counting Bloom Filters with single threaded and multithreaded insertion of k-mers of 50,000 long reads. The single threaded version does not suffer from race conditions and thus has the ground truth. The more threads we add the more visible is the effect of race conditions, but thanks to the mitigation mechanism, the differences are small. **B)** Wall-clock time it takes to read in and run a simulated workload of 2ms per read on 250,000 long reads. The blue data points use efficient sequence reading code, kseq [@kseq], surrounded by a naive implementation of a critical section,  while the orange data points use SeqReader. Unlike the critical section implementation which plateaus after a number of threads, SeqReader scales well and keeps benefitting even from higher number of threads.\label{fig:scalability}](scalability.png){ width=100% }
 
 # Acknowledgements
 
