### 2022-09-16 #4720 @vlad0x00
Hi!

I've added the code to reproduce the figures in the `paper` branch of btllib repo: https://github.com/bcgsc/btllib/tree/paper
The code can be found in `benchmarks` directory. It was tested under a Linux system and requires a recent `gcc` compiler, and `Python` and `R` to plot the figures. The code tests btllib with up to 128 threads so to properly reproduce the results your machine would need at least that many CPUs. If you're unable to reproduce up to that many threads, you can edit `benchmarks/run-test-cbf` and `benchmarks/run-test-input` scripts and remove the thread counts you don't want to test. To run the code, from within `benchmarks` dir, run `make`. This will run the code that generates the data and plots the `input_plot.png` (SeqReader vs kseq) and `cbf_plot.png` (CoutingBloomFilter) figures.
