### 2022-09-20 #4720 @vlad0x00
Just a quick comment -- I updated the scripts in `paper` branch to download the latest btllib version. The one it downloaded previously had some missing files which could've caused compile issues. If you successfully compiled it with the previous `paper` branch, that's alright, the code should've worked fine, it was just a compilation-related issue.
