### 2022-09-24 #4720 @pjb7687
I've run the benchmarks.

1. I had to change below in Makefile:

```
test_kseq: test_kseq.cpp
	g++ -std=c++17 -fopenmp -pthread -march=native -O3 -flto -lz -lm $< -o $@ -I/projects/btl/vnikolic/miniconda/include -L/projects/btl/vnikolic/miniconda/lib
```

to:

```
test_kseq: test_kseq.cpp
	g++ -std=c++17 -fopenmp -pthread -march=native -O3 -flto $< -lz -lm -o $@
```

so that it works properly. I've also installed all required Python and R packages.

2. Unfortunately, it takes forever for me to download `ERR6668574` from SRA, I had to run the benchmarks with a FASTQ file that I have.

3. I mainly focused on reproducing the second panel of the figure in the paper, which shows the performance improvement of `btllib` over `kseq`. Here is the result:

First round:
![input_plot](https://user-images.githubusercontent.com/1440536/192102927-de10cf9e-0acf-4784-8c2d-557c8d52bdb0.png)

Secound round:
![input_plot_2](https://user-images.githubusercontent.com/1440536/192104845-a2df5a66-c2b5-42c8-af26-3cd87fb9a0a5.png)

It seems that I couldn't reproduce the figure. I have run the benchmarks twice and have the same result.

4. The benchmarks seem to measure the total running time of the benchmark programs, not solely the I/O time taken by each algorithm. Since the kseq benchmark code is longer, and I suspect that it was the main reason the `btllib` was faster in the benchmark performed by the authors. I think in my case it was negligible since I use quite fast CPUs (2x Intel Xeon 5220R).

5. Finally I don't fully understand why it is important to test file I/O with such a lot of threads. In my opinion, usually file I/O is capped by the disk I/O speed not the number of threads.
