### 2022-09-24 adebd86
Remove misplaced include and library paths
