### 2022-09-24 #4720 @vlad0x00
Hi @pjb7687 ,

Thank you for additional feedback! To address your concerns:
> I had to change below in Makefile:

Thank you for spotting this, I've removed the misplaced include and library paths in the paper branch Makefile.

> Unfortunately, it takes forever for me to download ERR6668574 from SRA, I had to run the benchmarks with a FASTQ file that I have.

I've noticed that downloading sequences from there sometimes takes a bit and is fast at other times.

> It seems that I couldn't reproduce the figure. I have run the benchmarks twice and have the same resul

If your sequences aren't particularly long, the effects won't be visible since the time spent in the critical section performing I/O will be short. kseq is very efficient (hence why it's used for comparison) and so when working with short sequences kseq and SeqReader will perform about the same. Are you able to calculate N50 of the sequences and post it here? E.g. you can use [QUAST](https://quast.sourceforge.net/). The sequences used to generate the figure in the manuscript have an N50 of 41,459.

> The benchmarks seem to measure the total running time of the benchmark programs, not solely the I/O time taken by each algorithm. I noticed that the kseq benchmark code is longer, and I suspect that it was the main reason the btllib seemed to be faster in the benchmark performed by the authors. I think in my case it was negligible since I use quite fast CPUs (2x Intel Xeon 5220R).

SeqReader is likely the same as kseq I/O performance alone, since kseq is implemented very efficiently. What SeqReader does have is an efficient implementation of thread safety when multiple threads read from it. kseq is not threadsafe but default, which is why it needs a bit more code than SeqReader in order for multiple threads to be able to read from it. A modern bioinformatics application that does a lot of processing on large amounts of data is likely to benefit greatly from using multiple threads, especially if it wants to be scalable.

> Finally I don't fully understand why it is important to test file I/O with such a lot of threads. In my opinion, usually file I/O is capped by the disk I/O speed not the number of threads.

Often processing sequences can be parallelized at the level of each sequence, i.e. each sequence is assigned to a thread. A common pattern in my lab's code is to have multiple threads load sequences and then, e.g., store them in a Bloom filter or extract information that might be used for assembly or [scaffolding](https://en.wikipedia.org/wiki/Scaffolding_(bioinformatics)). SeqReader allows you to do this efficiently and maybe often more importantly with very little code.

I would like to add, however, that this performance part is not the most important part of SeqReader. Its most important benefit is ease of use (intuitive syntax and out-of-the-box thread safety), and flexibility (can read many compressed formats and multiple sequence formats (FASTA, FASTQ, SAM).
