### 2022-09-26 573aa5d
Preferably make pipes in /dev/shm/
