+++ src/btllib/process_pipeline.cpp
@@ -264,11 +264,12 @@ open_comm_pipes(
       if (access(pipepath.c_str(), F_OK) != -1) {
         unlink(pipepath.c_str());
       }
-      mkfifo(pipepath.c_str(), OPEN_MODE);
+      auto ret = mkfifo(pipepath.c_str(), OPEN_MODE);
+      check_error(ret != 0, "Process pipeline: mkfifo error: " + get_strerror());
 
       len = pipepath.size() + 1;
       check_error(len > COMM_BUFFER_SIZE,
-                  "Pipe path length too large for the buffer.");
+                  "Process pipeline: Pipe path length too large for the buffer.");
       check_error(!write_to_user(&len, sizeof(len)) ||
                     !write_to_user(pipepath.c_str(), len),
                   "Process pipeline: Communication failure.");
@@ -295,7 +296,8 @@ open_comm_pipes(
       const auto status_flags = fcntl(pipe_fd, F_GETFL);
       check_error(status_flags == -1,
                   "Process pipeline: fcntl error: " + get_strerror());
-      check_error(fcntl(pipe_fd, F_SETFL, status_flags & ~O_NONBLOCK) == -1,
+      ret = fcntl(pipe_fd, F_SETFL, status_flags & ~O_NONBLOCK);
+      check_error(ret == -1,
                   "Process pipeline: fcntl error: " + get_strerror());
       check_error(!write_to_user(&confirmation, sizeof(confirmation)),
                   "Process pipeline: Communication failure.");
@@ -311,23 +313,29 @@ static void
 redirect_io(const int in_fd, const int out_fd, const int err_fd)
 {
   if (in_fd != STDIN_FILENO) {
-    check_error(dup2(in_fd, STDIN_FILENO) == -1,
+    auto ret = dup2(in_fd, STDIN_FILENO);
+    check_error(ret == -1,
                 "Process pipeline: dup2 failed: " + get_strerror());
-    check_error(close(in_fd) != 0,
+    ret = close(in_fd);
+    check_error(ret != 0,
                 "Process pipeline: File descriptor close error: " +
                   get_strerror());
   }
   if (out_fd != STDOUT_FILENO) {
-    check_error(dup2(out_fd, STDOUT_FILENO) == -1,
+    auto ret = dup2(out_fd, STDOUT_FILENO);
+    check_error(ret == -1,
                 "Process pipeline: dup2 failed: " + get_strerror());
-    check_error(close(out_fd) != 0,
+    ret = close(out_fd); 
+    check_error(ret != 0,
                 "Process pipeline: File descriptor close error: " +
                   get_strerror());
   }
   if (err_fd != STDERR_FILENO) {
-    check_error(dup2(err_fd, STDERR_FILENO) == -1,
+    auto ret = dup2(err_fd, STDERR_FILENO);
+    check_error(ret == -1,
                 "Process pipeline: dup2 failed: " + get_strerror());
-    check_error(close(err_fd) != 0,
+    ret = close(err_fd);
+    check_error(ret != 0,
                 "Process pipeline: File descriptor close error: " +
                   get_strerror());
   }
@@ -594,10 +602,12 @@ run_cmd()
     if (idx > 0) {
       check_error(pipe(chainpipe_in_fd) == -1,
                   "Process pipeline: Error opening a pipe.");
-      check_error(fcntl(chainpipe_in_fd[PIPE_READ_END], F_SETFD, FD_CLOEXEC) ==
+      auto ret = fcntl(chainpipe_in_fd[PIPE_READ_END], F_SETFD, FD_CLOEXEC);
[…]
