### 2022-09-26 #4720 @MaximLippeveld
Thanks for uploading the code to reproduce the figures. I am able to reproduce figure 1A up to 64 threads. I don't have access to a bigger machine.

Figure 1B is for me also not showing a significant advantage for SeqReader, but I would like to rerun it to make sure. In any case, I think figure 1B should be added to the paper as it at least shows that SeqReader is on par with kseq, which is also good to know.

Since you mention performance is not the most important part, maybe you could also provide a code sample in the paper to show the intuitive syntax? @fboehm are code samples common in JOSS papers?
