### 2022-09-27 #4720 @pjb7687
> I've noticed that downloading sequences from there sometimes takes a bit and is fast at other times.

I tried it again yesterday and successfully downloaded the fastq file.

> If your sequences aren't particularly long, the effects won't be visible since the time spent in the critical section performing I/O will be short. kseq is very efficient (hence why it's used for comparison) and so when working with short sequences kseq and SeqReader will perform about the same. Are you able to calculate N50 of the sequences and post it here? E.g. you can use [QUAST](https://quast.sourceforge.net/). The sequences used to generate the figure in the manuscript have an N50 of 41,459.

Now I clearly understand the point. My fastq file only contains short reads as it is conventional Illumina sequencing data. I'll test it with the downloaded sequencing data.

> I would like to add, however, that the performance aspect is not the most important part of SeqReader. Its most important benefit is ease of use (intuitive syntax and out-of-the-box thread safety), and flexibility (can read many compressed formats and multiple sequence formats (FASTA, FASTQ, SAM). If you think this figure does not add much, I might as well remove it from the manuscript.

Now I believe that it is actually a pretty good point, and the figure should be included in the manuscript. I also would like to ask the authors to add just a short sentence that clearly describes this point, as SeqReader can be particularly useful when one has to deal with long-read data in many threads.

I will post again after I finish the test.
