### 2022-09-27 #4720 @fboehm
@MaximLippeveld - yes, short examples of code are fairly regular features in JOSS articles.
