### 2022-09-29 #4720 @pjb7687
![image](https://user-images.githubusercontent.com/1440536/193056616-d71d41b2-758d-4f67-bfca-af0ddaed7991.png)

Yes, now I can reproduce the figure.

As I mentioned in the above comment, I would like to kindly ask the authors to add just one sentence to clarify this. I do not have any further concerns.
