### 2022-09-30 dceb197
Add examples; clarify SeqReader's use case
