### 2022-09-30 #4720 @vlad0x00
Hi!

I've added an example of how to use SeqReader in the manuscript, as well as two more examples for ntHash and Counting Bloom Filter. Additionally, I've included a sentence that clarifies that SeqReader excels when loading long reads since they take more I/O time and consequently in the critical section as well.

The `paper` branch has the latest manuscript, and the following commit shows the changes I've made: https://github.com/bcgsc/btllib/commit/1ad3717c08b85ce26eb5a78b106320943525672a

Let me know if you require any other changes.

Cheers!
Vladimir
