### 2022-10-03 #4720 @MaximLippeveld
Thanks for adding the code samples. I have one final remark concerning the Design & Implementation section. It is not very clear to me if the listed algorithms and datastructures are reimplementations of existing software or if they are all original pieces of code. I assume the Bloom and Counting Bloom filter and Sequence I/O are original implementations, and ntHash, Multi-index bloom filter and Indexlr are reimplementations. Is this correct? If they are reimplementations, were the algorithms only ported from one language to another, or were they also optimized for scalability?
