### 2022-10-03 #4720 @vlad0x00
Except for sequence I/O, all of the code in btllib started out in different projects in our lab. To make it easier to reuse, we moved the commonly used code into btllib which now has the reference/latest implementations for the listed algorithms. The code wasn't just moved, though, in general most of it had some improvements in efficiency, portability, or readability. For some of the algorithms/data structures, the original project is cited in btllib, whereas for others that weren't published it's not. E.g. counting bloom filter wasn't published, our lab simply had it in its own repository until it was moved to btllib. Let me know if you need further explanation or if I should modify the manuscript text to reflect any of this.
