### 2022-10-05 #4720 @MaximLippeveld
Ok, thanks for clarifying. Yes, it would be nice if this is reflected in the manuscript. Perhaps you could add in the Design & Implementation section that implementations for ntHash, Multi-index Bloom filter and Indexlr were adapted from previously existing work. Other than that, I have no further remarks!
