### 2022-10-20 b44daa8
Add deletion to CBF
