### 2022-10-20 56a8497
Add clear function to CBF
