### 2022-10-20 7f5a417
Fix deletion infinity loop
