### 2022-10-24 #4720 @fboehm
@vlad0x00 - the paper looks great!

My only request is that you add a doi for the Reinert article. When I follow the hyperlink that you've provided, it directs me here: http://publications.imp.fu-berlin.de/2103/

I think that you need 10.1016/j.jbiotec.2017.07.017 as the doi, but please check this.

Thanks again!
