+++ paper.bib
@@ -21,7 +21,8 @@ NGS analysis; Software libraries; C++; Data structures},
  pages = {157--168},
  publisher = {ELSEVIER},
  title = {The SeqAn C++ template library for efficient sequence analysis: A resource for programmers},
- url = {http://publications.imp.fu-berlin.de/2103/},
+ url = {http://doi.org/10.1016/j.jbiotec.2017.07.017},
+ doi = {10.1016/j.jbiotec.2017.07.017},
  volume = {261},
  year = {2017}
 }
