### 2022-10-25 #4720 @vlad0x00
Thank you, @fboehm !

I've added the doi: https://github.com/bcgsc/btllib/commit/7be06f88097099afd426494f8c8ed241281f2a97
