### 2022-10-27 3b6f139
Change clang-format source for MacOS CI
