### 2022-10-27 dfc9412
Fix clang-tidy complaints
