+++ src/btllib/seq_reader_sam_module.cpp
@@ -93,7 +93,7 @@ SeqReaderSamModule::buffer_valid(const char* buffer, const size_t size)
           }
           break;
         case TLEN:
-          if (!(bool(std::isdigit(c)) || c == '-')) {
+          if (!bool(std::isdigit(c)) && c != '-') {
             return false;
           }
           break;
