### 2022-10-27 d19b770
Add thread-unsafe update function
