### 2022-10-28 #4720 @vlad0x00
Done. I've archived the `paper` branch which has the version of btllib at the time of the writing of the manuscript.

https://doi.org/10.5281/zenodo.7259920
