### 2022-10-28 #4720 @vlad0x00
Here's the updated URLs:

Release: https://github.com/bcgsc/btllib/releases/tag/joss-release2
DOI: https://doi.org/10.5281/zenodo.7259920
