### 2022-10-28 #4720 @danielskatz
Hi - I just happened to see this thread - For JOSS, the paper source does not *need* to be in the archived software repo; the archive can just be of the main branch. The final paper itself is archived separately in any case.
