### 2022-10-28 #4720 @warrenlr
Hello everyone, thank you for all your hard work on this.
Just curious, do we have an ETA date for the release of the DOI for the btllib manuscript?  I am asking because we'd like to include a citation in a research proposal and publication submission to Nat. met. due Monday Oct. 31st.
Thank you,
René
