### 2022-10-28 #4720 @vlad0x00
@fboehm Do you want me to make a release without the manuscript in it?
