### 2022-10-28 #4720 @fboehm
I can confirm that the pdf authors match exactly those specified in the archive.
