### 2022-10-28 #4720 @fboehm
> Hello everyone, thank you for all your hard work on this. Just curious, do we have an ETA date for the release of the DOI for the btllib manuscript? I am asking because we'd like to include a citation in a research proposal and publication submission to Nat. met. due Monday Oct. 31st. Thank you, René

Hi, @warrenlr - once the changes to the archive title are made, I will signal to the editors in chief that the manuscript is ready for publication. Once they have a chance to approve it, the doi will be issued. I can't say when exactly that will occur; it depends on their schedules.
