### 2022-11-01 dfa1341
Fix CBF insertion
