### 2022-11-02 #4720 @Kevin-Mattheus-Moerman
@vlad0x00 I have read your paper and reviewed the ZENODO archive. Most seems in order, I just have the following minor comments:

- [ ] Please manually edit the ZENODO archive so that the version tag matches that of the paper and software i.e. `joss-release2`, I believe it currently says version 2.
- [ ] Please also manually edit the ZENODO archive license. It currently lists a Creative Commons license but it should be your software license.
- [ ] Recommendation: You may remove the zip-codes from you affiliations.
- [ ] Reading this sentence I felt the negative "not well" did not distribute well: `This could be because the pre-existing computer programming code is either not well written, widely available, documented, efficient nor general enough.` Could something like this be an improvement: `This could be because the pre-existing computer programming code is either not well written, not widely available, insufficiently documented, inefficient, or not general enough.`? Feel free to disagree or propose something else.
