### 2022-11-03 #4720 @vlad0x00
Hi @Kevin-Mattheus-Moerman ,

I've removed the zip codes and updated the sentence you mentioned, as can be seen in this commit: https://github.com/bcgsc/btllib/commit/a17c760feb9524c908155ae3740f186f85fccab9

Since the paper has been updated, and the previous Zenodo archive contains the outdated paper, I've instead released a new archive with just the master branch v1.4.9 release:
Release: https://github.com/bcgsc/btllib/releases/tag/v1.4.9
DOI: https://doi.org/10.5281/zenodo.7275416

The licence has also been changed to reflect the repository licence.
