### 2022-11-03 #4720 @Kevin-Mattheus-Moerman
@vlad0x00 the reviewed the changes to the paper and all is good. However the archive and version tag still need your attention.

The normal process is for the JOSS publication to be associated with a particular version of your software. Then this version should also appear on github as a tagged version. Next this is also the version listed on ZENODO. Currently your most recent version tag which seems associated with this review is `joss-release2`. However the version tag you asked me to assign for this review now reads `v1.4.9` (which seems to relate to an older Sept release on GitHub), and the ZENDODO archive says `Version btllib v1.4.9`. We need all of these to be the exact same and the name should correspond to a tagged release on GitHub. Can you make this happen? Please inform me if we need to make changes to the version tag here or if we need to update the ZENODO archive again.
If you need to create a new release perhaps follow these automated steps: https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content which will create a ZENODO entry automatically, that should populate most fields correctly automatically too (however you probably do need to manually edit the title , author list, and license if you choose this approach).
Let me know what approach you take.
