### 2022-11-04 #4720 @vlad0x00
The v1.4.9 version is the latest release and reflects the functionality described in the paper, and has the performance characteristics tested by the reviewers, so I think it's a sensible version to use for JOSS. I can make a new release, but it'd basically be identical to v1.4.9.

I've updated the Zenodo archive (https://zenodo.org/record/7275416) version to be named "v1.4.9". It reflects the GitHub tag v1.4.9 (https://github.com/bcgsc/btllib/tree/v1.4.9).
