### 2022-11-04 #4720 @vlad0x00
The v1.4.9 version is the currently latest release and reflects the functionality described in the paper, and has the performance characteristics tested by the reviewers, so I think it's a sensible version to use for JOSS. `joss-release` and `joss-release2` were functionally identical to v1.4.9 but were on the `paper` branch instead, which contains the manuscript.

I've updated the Zenodo archive (https://zenodo.org/record/7275416) version to be named "v1.4.9". It reflects the GitHub tag v1.4.9 (https://github.com/bcgsc/btllib/tree/v1.4.9). Let me know if there's any other steps I missed.

I've also updated the funding section in the paper to add a funder that should be there, as advised by one of the senior authors, @warrenlr. The changes can be seen in https://github.com/bcgsc/btllib/commit/49954f53fc98f8a24fd90150c697090903b5d732 and https://github.com/bcgsc/btllib/commit/5e7eb7749dad711fb29f10921d87e69c2e079e9e.

Cheers
